feat: update import paths for InfraredVaultContractABI and reorganize file structure
